backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.0.0
pytest>=7.4.0
pytest-cov>=4.1.0
httpx>=0.25.0

# LinkedIn Automation V1 dependencies
gspread>=5.12.0
google-auth>=2.23.0
playwright>=1.40.0
python-dotenv>=1.0.0

